requirements.txt
Django==4.0.4
django-phonenumber-field==6.1.0
phonenumbers==8.12.47
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6